#49 Refactor contributor modal with RmgSection component
