+++ package.json
@@ -10,7 +10,7 @@
         "@chakra-ui/react": "^2.8.1",
         "@emotion/react": "^11.11.1",
         "@emotion/styled": "^11.11.0",
-        "@railmapgen/rmg-components": "^8.2.2",
+        "@railmapgen/rmg-components": "^9.0.0",
         "@railmapgen/rmg-runtime": "^8.0.3",
         "@railmapgen/rmg-translate": "^3.1.0",
         "@reduxjs/toolkit": "^1.9.5",
@@ -25,25 +25,25 @@
     "devDependencies": {
         "@testing-library/jest-dom": "^6.1.3",
         "@testing-library/react": "^14.0.0",
-        "@types/node": "^18.17.17",
-        "@types/react": "^18.2.21",
+        "@types/node": "^18.17.19",
+        "@types/react": "^18.2.22",
         "@types/react-dom": "^18.2.7",
         "@types/react-router-dom": "^5.3.3",
         "@types/redux-mock-store": "^1.0.3",
-        "@typescript-eslint/eslint-plugin": "^6.7.0",
-        "@typescript-eslint/parser": "^6.7.0",
+        "@typescript-eslint/eslint-plugin": "^6.7.2",
+        "@typescript-eslint/parser": "^6.7.2",
         "@vitejs/plugin-legacy": "^4.1.1",
         "@vitejs/plugin-react": "^4.0.4",
-        "eslint": "^8.49.0",
+        "eslint": "^8.50.0",
         "eslint-plugin-prettier": "^5.0.0",
         "eslint-plugin-react": "^7.33.2",
         "jsdom": "^22.1.0",
         "prettier": "^3.0.3",
         "redux-mock-store": "^1.5.4",
-        "terser": "^5.19.4",
+        "terser": "^5.20.0",
         "typescript": "^5.2.2",
         "vite": "^4.4.9",
-        "vitest": "^0.34.4"
+        "vitest": "^0.34.5"
     },
     "scripts": {
         "dev": "vite --host",
